requirements.txt
asgiref>=3.5.2
autopep8>=2.0.1
django	
django-active-link>=0.1.8
django-ckeditor
django-js-asset>=2.0.0
fontawesome>=5.10.1.post1
fontawesomefree>=6.3.0
Pillow>=9.1.1
pycodestyle>=2.10.0
pytz>=2022.1
sqlparse>=0.4.2
tomli>=2.0.1
tzdata>=2022.1
whitenoise
mysqlclient
